Pull item meta from itemdb class
