+++ Essentials/src/com/earth2me/essentials/ItemDb.java
@@ -5,9 +5,7 @@ import com.earth2me.essentials.api.IItemDb;
 import java.util.*;
 import java.util.regex.Pattern;
 import org.bukkit.Material;
-import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.meta.EnchantmentStorageMeta;
 
 
 public class ItemDb implements IConf, IItemDb
@@ -139,95 +137,7 @@ public class ItemDb implements IConf, IItemDb
 		retval.setDurability(metaData);
 		return retval;
 	}
-
-	public void addStringEnchantment(final User user, final boolean allowUnsafe, final ItemStack stack, final String string) throws Exception
-	{
-		final String[] split = splitPattern.split(string, 2);
-		if (split.length < 1)
-		{
-			return;
-		}
-
-		Enchantment enchantment = getEnchantment(user, split[0]);
-
-		int level = -1;
-		if (split.length > 1)
-		{
-			try
-			{
-				level = Integer.parseInt(split[1]);
-			}
-			catch (NumberFormatException ex)
-			{
-				level = -1;
-			}
-		}
-
-		if (level < 0 || (!allowUnsafe && level > enchantment.getMaxLevel()))
-		{
-			level = enchantment.getMaxLevel();
-		}
-		addEnchantment(user, allowUnsafe, stack, enchantment, level);
-	}
-
-	public void addEnchantment(final User user, final boolean allowUnsafe, final ItemStack stack, final Enchantment enchantment, final int level) throws Exception
-	{
-		try
-		{
-			if (stack.getType().equals(Material.ENCHANTED_BOOK))
-			{
-				EnchantmentStorageMeta meta = (EnchantmentStorageMeta)stack.getItemMeta();
-				if (level == 0)
-				{
-					meta.removeStoredEnchant(enchantment);
-				}
-				else
-				{
-					meta.addStoredEnchant(enchantment, level, allowUnsafe);
-				}
-				stack.setItemMeta(meta);
-			}
-			else // all other material types besides ENCHANTED_BOOK
-			{
-				if (level == 0)
-				{
-					stack.removeEnchantment(enchantment);
-				}
-				else
-				{
-					if (allowUnsafe)
-					{
-						stack.addUnsafeEnchantment(enchantment, level);
-					}
-					else
-					{
-						stack.addEnchantment(enchantment, level);
-					}
-				}
-			}
-		}
-		catch (Exception ex)
-		{
-			throw new Exception("Enchantment " + enchantment.getName() + ": " + ex.getMessage(), ex);
-		}
-	}
-
-	//TODO: Properly TL this
-	public Enchantment getEnchantment(final User user, final String name) throws Exception
-	{
-		final Enchantment enchantment = Enchantments.getByName(name);
-		if (enchantment == null)
-		{
-			throw new Exception(_("enchantmentNotFound") + ": " + name);
-		}
-		final String enchantmentName = enchantment.getName().toLowerCase(Locale.ENGLISH);
-		if (user != null && !user.isAuthorized("essentials.enchant." + enchantmentName))
-		{
-			throw new Exception(_("enchantmentPerm", enchantmentName));
-		}
-		return enchantment;
-	}
-
+	
 	public String names(ItemStack item)
 	{
 		ItemData itemData = new ItemData(item.getTypeId(), item.getDurability());
