+++ Essentials/src/com/earth2me/essentials/Kit.java
@@ -115,24 +115,25 @@ public class Kit
 				}
 
 				final String[] parts = d.split(" ");
-				final ItemStack stack = ess.getItemDb().get(parts[0], parts.length > 1 ? Integer.parseInt(parts[1]) : 1);
+				final ItemStack parseStack = ess.getItemDb().get(parts[0], parts.length > 1 ? Integer.parseInt(parts[1]) : 1);
+				final MetaItemStack stack = new MetaItemStack(parseStack);
 
 				if (parts.length > 2)
 				{
 					for (int i = 2; i < parts.length; i++)
 					{
-						ess.getItemDb().addStringEnchantment(null, allowUnsafe, stack, parts[i]);
+						stack.addStringEnchantment(null, allowUnsafe, parts[i]);
 					}
 				}
 
 				final Map<Integer, ItemStack> overfilled;
 				if (user.isAuthorized("essentials.oversizedstacks"))
 				{
-					overfilled = InventoryWorkaround.addOversizedItems(user.getInventory(), ess.getSettings().getOversizedStackSize(), stack);
+					overfilled = InventoryWorkaround.addOversizedItems(user.getInventory(), ess.getSettings().getOversizedStackSize(), stack.getBase());
 				}
 				else
 				{
-					overfilled = InventoryWorkaround.addItems(user.getInventory(), stack);
+					overfilled = InventoryWorkaround.addItems(user.getInventory(), stack.getBase());
 				}
 				for (ItemStack itemStack : overfilled.values())
 				{
