+++ Essentials/src/com/earth2me/essentials/commands/Commandenchant.java
@@ -2,6 +2,7 @@ package com.earth2me.essentials.commands;
 
 import com.earth2me.essentials.Enchantments;
 import static com.earth2me.essentials.I18n._;
+import com.earth2me.essentials.MetaItemStack;
 import com.earth2me.essentials.User;
 import com.earth2me.essentials.Util;
 import java.util.Locale;
@@ -10,7 +11,6 @@ import java.util.Set;
 import java.util.TreeSet;
 import org.bukkit.Server;
 import org.bukkit.enchantments.Enchantment;
-import org.bukkit.inventory.ItemStack;
 
 
 public class Commandenchant extends EssentialsCommand
@@ -24,7 +24,7 @@ public class Commandenchant extends EssentialsCommand
 	@Override
 	protected void run(final Server server, final User user, final String commandLabel, final String[] args) throws Exception
 	{
-		final ItemStack stack = user.getItemInHand();
+		final MetaItemStack stack = new MetaItemStack(user.getItemInHand());
 		if (stack == null)
 		{
 			throw new Exception(_("nothingInHand"));
@@ -58,11 +58,11 @@ public class Commandenchant extends EssentialsCommand
 		}
 
 		final boolean allowUnsafe = ess.getSettings().allowUnsafeEnchantments() && user.isAuthorized("essentials.enchant.allowunsafe");
-		final Enchantment enchantment = ess.getItemDb().getEnchantment(user, args[0]);
-		ess.getItemDb().addEnchantment(user, allowUnsafe, stack, enchantment, level);
+		final Enchantment enchantment = stack.getEnchantment(user, args[0]);
+		stack.addEnchantment(user, allowUnsafe, enchantment, level);
 
 
-		user.getInventory().setItemInHand(stack);
+		user.getInventory().setItemInHand(stack.getBase());
 		user.updateInventory();
 		final String enchantmentName = enchantment.getName().toLowerCase(Locale.ENGLISH);
 		if (level == 0)
